pyproject.toml
[tool.poetry]
name = "open-ai"
version = "0.1.0"
description = ""
authors = ["Alisson <alissonpintor@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.40.8"
ipykernel = "^6.29.5"
python-dotenv = "^1.0.1"
yfinance = "^0.2.41"
pandas = "^2.2.2"
matplotlib = "^3.9.2"
pywa = "^1.25.0"
pyaudio = "^0.2.14"
speechrecognition = "^3.10.4"
sounddevice = "^0.5.0"
playsound = "1.2.2"
setuptools = "^73.0.1"
python-fasthtml = "^0.4.5"
transformers = "^4.44.2"
numpy = "<2"
pybind11 = ">=2.12"
pdf2image = "^1.17.0"
streamlit = "^1.37.1"
mypy = "^1.11.2"
pyright = "^1.1.377"
streamlit-cropper = "^0.2.2"
tabula-py = "^2.9.3"


[tool.poetry.group.dev.dependencies]
mypy = "^1.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
